feat: use same account component
