--- cross-app-connect/src/app/components/AccountChip.module.css
@@ -1,146 +1,0 @@
-.chip {
-    display: flex;
-    align-items: center;
-    gap: 12px;
-    padding: 12px;
-    border-radius: var(--radius-md);
-    background: var(--bg-card-elevated);
-    border: 1px solid var(--border-default);
-}
-
-.avatar {
-    width: 36px;
-    height: 36px;
-    border-radius: var(--radius-full);
-    flex-shrink: 0;
-    background: var(--login-btn-hover-bg);
-}
-
-.avatarPlaceholder {
-    width: 36px;
-    height: 36px;
-    border-radius: var(--radius-full);
-    background: var(--login-btn-hover-bg);
-    flex-shrink: 0;
-}
-
-.identity {
-    display: flex;
-    flex-direction: column;
-    flex: 1;
-    min-width: 0;
-    gap: 0;
-}
-
-.label {
-    font-size: 12px;
-    color: var(--text-subtle);
-    margin: 0;
-}
-
-.identityRow {
-    display: flex;
-    align-items: center;
-    gap: 8px;
-}
-
-.domainStack {
-    display: flex;
-    flex-direction: column;
-    min-width: 0;
-}
-
-.domain {
-    font-weight: 600;
-    color: var(--text-strong);
-    line-height: 1.2;
-    margin: 0;
-}
-
-.addressUnderDomain {
-    font-family: ui-monospace, 'SF Mono', Menlo, Consolas, monospace;
-    font-size: 12px;
-    color: var(--text-subtle);
-    line-height: 1.2;
-    margin: 0;
-}
-
-.addressOnly {
-    font-family: ui-monospace, 'SF Mono', Menlo, Consolas, monospace;
-    font-size: 14px;
-    font-weight: 500;
-    color: var(--text-strong);
-    margin: 0;
-}
-
-.copyBtn {
-    display: inline-flex;
-    align-items: center;
-    justify-content: center;
-    width: 22px;
-    height: 22px;
-    min-width: 22px;
-    background: transparent;
-    border: none;
-    border-radius: var(--radius-sm);
-    color: var(--text-muted);
-    padding: 0;
-}
-
-.copyBtn:hover {
-    background: var(--login-btn-hover-bg);
-}
-
-.copyIcon {
-    width: 14px;
-    height: 14px;
-}
-
-.balanceCol {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-end;
-    gap: 0;
-}
-
-.balanceList {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-end;
-    gap: 0;
-}
-
-.balanceRow {
-    font-size: 14px;
-    font-weight: 600;
-    color: var(--text-strong);
-    font-family: ui-monospace, 'SF Mono', Menlo, Consolas, monospace;
-    margin: 0;
-}
-
-.skeleton {
-    background: var(--login-btn-hover-bg);
-    border-radius: var(--radius-sm);
-    animation: pulse 1.5s ease-in-out infinite;
-}
-
-.skeletonAvatar {
-    width: 36px;
-    height: 36px;
-    border-radius: var(--radius-full);
-}
-
-.skeletonText {
-    height: 14px;
-    width: 70px;
-}
-
-.skeletonDomain {
-    height: 16px;
-    width: 110px;
-}
-
-@keyframes pulse {
-    0%, 100% { opacity: 1; }
-    50% { opacity: 0.55; }
-}
